+++ src/pc/utils.js
@@ -3,12 +3,14 @@ const childProcess = require('child_process')
 const fs = require('fs')
 const path = require('path')
 
+const { PLATFORM } = require('./constants')
+
 const log = (...args) => console.log(args.join(' '))
 const logDebug = (...args) => process.env.NODE_ENV === 'development' ? log(chalk.gray(...args)) : null
 const logError = (...args) => console.log(chalk.redBright(args.join(' ')))
 
 const open = (url) => {
-  const start = (process.platform === 'darwin' ? 'open' : process.platform === 'win32' ? 'start' : 'xdg-open')
+  const start = (PLATFORM === 'darwin' ? 'open' : PLATFORM === 'win32' ? 'start' : 'xdg-open')
   childProcess.exec(start + ' ' + url)
 }
 
@@ -36,10 +38,15 @@ const copyRecursiveSync = (src, dest) => {
   }
 }
 
+const isRoot = () => {
+  return PLATFORM === 'darwin' ? process.getuid && process.getuid() === 0 : false
+}
+
 module.exports = {
   log,
   logDebug,
   logError,
   open,
-  copyRecursiveSync
+  copyRecursiveSync,
+  isRoot
 }